Move the api-requests directory into common.
